# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Security scanning platform — manages targets, scan jobs, reports, and triage rules.",
        "title": "NexusSec API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/scans/{id}/report": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy báo cáo lỗ hổng của một scan đã hoàn thành. Kiểm tra quyền sở hữu qua PG rồi lấy dữ liệu từ MongoDB.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Get scan report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan Job ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_gateway_handler.ReportResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Scan chưa hoàn thành",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Scan hoặc report không tồn tại",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_nexussec_nexussec_pkg_response.JSONResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_gateway_handler.CleanVulnerability": {
            "type": "object",
            "properties": {
                "cvss_score": {
                    "type": "number"
                },
                "cwe": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "remediation": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "source_tool": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "vuln_id": {
                    "type": "string"
                }
            }
        },
        "internal_gateway_handler.ReportResponse": {
            "type": "object",
            "properties": {
                "scan_id": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/internal_gateway_handler.SummaryResponse"
                },
                "vulnerabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_gateway_handler.CleanVulnerability"
                    }
                }
            }
        },
        "internal_gateway_handler.SummaryResponse": {
            "type": "object",
            "properties": {
                "critical": {
                    "type": "integer"
                },
                "high": {
                    "type": "integer"
                },
                "info": {
                    "type": "integer"
                },
                "low": {
                    "type": "integer"
                },
                "medium": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Nhập: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}